requirements.txt
discord.py
aiofiles
feedparser
pytz
beautifulsoup4
requests
